Add support for WMQ V8 Tested on Windows 64 bit against MQ V7.5 and V8
