--- nbproject/project.properties
@@ -1,11 +1,0 @@
-examples.dir=${file.reference.trunk-examples}
-file.reference.trunk-examples=examples
-file.reference.trunk-ext=ext
-file.reference.trunk-lib=lib
-file.reference.trunk-tests=tests
-javac.classpath=
-lib.dir=${file.reference.trunk-lib}
-main.file=
-source.encoding=UTF-8
-src.dir=${file.reference.trunk-ext}
-test.src.dir=${file.reference.trunk-tests}
